Fix HDFS sink binding
 - Move HdfsSink configuration into the configuration class that is actually invoked by the spring boot main class

This resolves #27
