+++ hdfs/spring-cloud-starter-stream-sink-hdfs/src/test/java/org/springframework/cloud/stream/app/hdfs/sink/HdfsSinkConfigurationTests.java
@@ -60,7 +60,7 @@ public class HdfsSinkConfigurationTests {
 	private FsShell fsShell;
 
 	@Autowired
-	@Bindings(HdfsSink.class)
+	@Bindings(HdfsSinkConfiguration.class)
 	private Sink sink;
 
 	@BeforeClass
