+++ hdfs/spring-cloud-starter-stream-sink-hdfs/src/test/java/org/springframework/cloud/stream/app/hdfs/sink/HdfsSinkIntegrationTests.java
@@ -64,7 +64,7 @@ public class HdfsSinkIntegrationTests {
 	private FsShell fsShell;
 
 	@Autowired
-	@Bindings(HdfsSink.class)
+	@Bindings(HdfsSinkConfiguration.class)
 	private Sink sink;
 
 	@BeforeClass
